syn, ant and note in JSON, bold the word in example, README JSON instructions, improvements to split_from_disk.py
BIG COMMIT:
1.Added syn and ant entries in JSON. Modified the whole database, including
orginal-data and data. Added the relevant code in txt2json.py, all html
templates and in common.js so that it sends correct data to the server

2. Added the note field in the JSON sent by jquery to Flask. This would
have cause trouble otherwise. Also added display for that field in word.html

3. Made a regex that finds how the word is used in the example and boldens it
in JS. This is added to word.js

4. Added all you need to do to change the JSON database format as instructions
in README

5. Fixed the problem of default words not having their own list. Made the changes
in log.md

6. Added more automation to split_from_disk.py. Now it automatically overwrites
data/all.json. NOTE: THIS PART NEEDS FIXING IN THE FUTURE WHEN YOU HAVE MANUALLY
CHANGED MEANINGS AND YOU EXECUTE this script